Use css vars
I could merge light and dark css into one and just change the variables with JavaScript. But this would cause the page to render with a default light/dark theme and we would have to wait for JS to run and switch the theme. I don't like to keep two files, but having a page switch from light to dark or vice versa on render is too bad.